backend/requirements.txt
# Philadelphia Calendar Backend - Python Dependencies

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# HTML Parsing
html5lib==1.1

# Date Parsing
python-dateutil==2.8.2

# Scheduling
apscheduler==3.10.4

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Production WSGI server
gunicorn==21.2.0

# Environment Variables
python-dotenv==1.0.0

# Optional: For more advanced scraping
# selenium==4.15.2  # Uncomment if you need JavaScript rendering
# playwright==1.40.0  # Alternative to Selenium
